requirements.txt
numpy==1.16.2
pandas==0.24.2
scikit-image==0.16.2
scipy==1.2.1
tensorflow-gpu==2.6.0
tensorboard==2.6.0
matplotlib==3.1.0
jupyter==1.0.0
scikit-learn==0.20.3
progressbar==2.5
